refactor: replaced react-query custom hooks to components, deleted unusing components, moved Modal components out
