--- src/components/Presentation/ObjectEditor/AnimationEditor/EffectEditor/index.jsx
@@ -1,7 +1,0 @@
-function EffectEditor() {
-  return (
-    <div>EffectEditor</div>
-  );
-}
-
-export default EffectEditor;
