--- src/components/Presentation/ObjectEditor/AnimationEditor/OrderEditor/index.jsx
@@ -1,11 +1,0 @@
-function OrderEditor() {
-  return (
-    <ol>
-      <li>a</li>
-      <li>b</li>
-      <li>c</li>
-    </ol>
-  );
-}
-
-export default OrderEditor;
