--- src/components/Presentation/ObjectEditor/StyleEditor/ArrangeEditor/index.jsx
@@ -1,12 +1,0 @@
-function ArrangeEditor() {
-  return (
-    <ul>
-      <li>1</li>
-      <li>2</li>
-      <li>3</li>
-      <li>4</li>
-    </ul>
-  );
-}
-
-export default ArrangeEditor;
