--- src/components/Presentation/PresentationCanvas/index.jsx
@@ -1,5 +1,0 @@
-function PresentationCanvas() {
-
-}
-
-export default PresentationCanvas;
